fix(doctor): make disk-space check build on Darwin
syscall.Statfs_t.Bsize is int64 on Linux but uint32 on Darwin, so the
free-space multiplication `int64(st.Bavail) * st.Bsize` failed to compile on
macOS — the migration never built there. Cast both operands to int64.

Caught by the macOS leg of the CI build matrix.

+++ cmd/dots/commands.go
@@ -835,7 +835,9 @@ func runDoctor() int {
 	// Disk space
 	var st syscall.Statfs_t
 	if err := syscall.Statfs(home, &st); err == nil {
-		freeMB := int64(st.Bavail) * st.Bsize / (1024 * 1024)
+		// Bsize is int64 on Linux but uint32 on Darwin; cast both operands so the
+		// multiplication compiles on every platform.
+		freeMB := int64(st.Bavail) * int64(st.Bsize) / (1024 * 1024)
 		if freeMB < 100 {
 			warn(fmt.Sprintf("Low disk space in $HOME: %d MB free", freeMB))
 		} else {
